symbols: Add missing Build-Depends-Package fields
Add the Build-Depends-Package field to the symbols files so that
dpkg-shlibdeps can use it to ensure that the resulting binary package
dependency on the shared library is at least as strict as the source
package dependency on the shared library development package.

This fixes the symbols-file-missing-build-depends-package-field lintian
warnings. See also Debian Policy Section 8.6.3.2.