symbols, lintian: Drop reference to private symbols
In upstream commit 56c4857f1763 all the abi16_* functions and the
nouveau_debug variable got hidden, so the reference to the private
symbols can be dropped and also the lintian-override for it.

This did cause/show a invalid-template-id-in-symbols-file lintian error,
fix that by updating the version which added that function (publicly).
This in effect completes the 'revert' of commit f1c5b014206e.